backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
python-multipart==0.0.6
groq==0.4.1
python-dotenv==1.0.0
email-validator==2.1.0
google-generativeai==0.3.5
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
python-http-client==3.3.7